requirements.txt
streamlit
streamlit-option-menu
python-dotenv
google-generativeai
pypdf
langchain_google_genai
langchain_text_splitters
langchain_community
faiss-cpu
